[desktop] unboxing and loop style fixes

+++ desktop/src/org/gudy/azureus2/core3/util/Average.java
@@ -92,7 +92,7 @@ public class Average {
 	  Object[] details = other.getCloneDetails();
   		
   	  values		= (long[])details[0];
-  	  lastUpdate	= ((Long)details[1]).longValue();
+  	  lastUpdate	= (Long) details[1];
   }
   
   private synchronized Object[]
